social_crawler_demo/pyproject.toml
[project]
name = "social-crawler-demo"
version = "0.1.0"
description = "Capability demo: Scrapy + Playwright CDP attach architecture for public social media data (FB/Twitter/IG)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "scrapy>=2.11",
    "scrapy-playwright>=0.0.34",
    "playwright>=1.45",
    "python-dotenv>=1.0",
    "itemadapter>=0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["demo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
